requirements.txt
# Core Dependencies
PyPDF2==3.0.1
Pillow==10.0.0
python-dotenv==1.0.0
moviepy==1.0.3
tqdm==4.66.1
python-magic==0.4.27
watchdog==3.0.0
